pyproject.toml
[project]
name = "PowerControllerViewer"
version = "2.0.1"
description = "View the status of one or more PowerController or LightingControl application instances"
readme = "README.md"
requires-python = ">=3.13"
service_name="PowerControllerViewer"
launch_path="src/main.py"
dependencies = [
    "cerberus>=1.3.7",
    "fastapi>=0.115.0",
    "jinja2>=3.1.0",
    "org-enums",
    "pre-commit>=4.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sc-foundation-services>=3.0.4",
    "uvicorn[standard]>=0.30.0",
    "wheel>=0.45.1",
]

[tool.uv.sources]
org-enums = { git = "https://github.com/NickElseySpelloC/org_enums", rev = "main" }

[dev.env.sync]
remote="~/Library/CloudStorage/Dropbox/Development/dev_setup/app_config/PowerControllerViewer"
patterns = [
    "config.yaml",
    "state_data/*",
    "logs/*",
]
